Trait tartan_c_enum::CEnum

pub trait CEnum<T> where
    Self: Debug + Default + Clone + Copy + PartialEq + Hash + PartialOrd + Ord + From<T>,
    T: From<Self>, {
    fn name(self) -> Option<&'static str>;
}

Trait implemented by all c_enum types.

This mainly exists for documentation, to show which traits and methods are available on all C-style enums defined with this crate. (Click “Show Declaration” in Rustdoc to see the supertraits.)

Required Methods

fn name(self) -> Option<&'static str>

The name of the enum variant in code, if one is defined for this value.